pyproject.toml
[tool.poetry]
name = "mayuri"
version = "3.0.0"
description = "Modular Telegram Bot based on PyroFork"
license = "GPL-3.0-or-later"
authors = ["wulan17 <wulan17@nusantararom.org>"]
readme = "README.md"
repository = "https://github.com/Mayuri-Chan/Mayuri"

[tool.poetry.dependencies]
python = "~=3.9"
async-pymongo = ">=0.1.4,<1.0.0"
uvloop = ">=0.19.0,<1.0.0"
python-dotenv = "^1.0.0"
TgCrypto-pyrofork = ">=1.2.7,<2.0.0"
httpx = {extras = ["http2"], version = ">=0.26.0,<1.0.0"}
colorlog = ">=6.8.0,<7.0.0"
meval = ">=2.5,<3.0"
apscheduler = {git = "https://github.com/agronholm/apscheduler.git", rev = "master"}
pyrofork = ">=2.3.4,<2.4.0"
tomli = ">=2.0.1,<3.0.0"
opencv-python = ">=4.10.0.84,<5.0.0.0"
unidecode = ">=1.3.7,<2.0.0"
pytesseract = ">=0.3.10,<1.0.0"
pynacl = ">=1.5.0,<2.0.0"
google-genai = ">=1.0.0,<2.0.0"
aio-telegraph = ">=1.0.0,<2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
